+++ packages/dart_edge_core/lib/src/http.dart
@@ -1,5 +1,6 @@
 export 'http/binary_stream_response.dart';
 export 'http/error_response.dart';
+export 'http/http_byte_range.dart';
 export 'http/http_method.dart';
 export 'http/json_decoding.dart';
 export 'http/multipart_form_data.dart';